I have drawable points now, but I'm running into a small issue.  It seems that the grid layout's geometry isn't matching the bounding box I'm setting.  I set a 350 x 350 view for the points to be printed in, but they only print within a small subrectangle of the grid layout...
Also, made the point list a singleton class.  No destructor or clear function yet.  To come later.

+++ point.cpp
@@ -6,28 +6,28 @@ Point::Point()
     this->y = 0;
 }
 
-Point::Point(int x, int y)
+Point::Point(double x, double y)
 {
     this->x = x;
     this->y = y;
 }
 
-int Point::getX()
+double Point::getX()
 {
     return x;
 }
 
-int Point::getY()
+double Point::getY()
 {
     return y;
 }
 
-void Point::setX(int x)
+void Point::setX(double x)
 {
     this->x = x;
 }
 
-void Point::setY(int y)
+void Point::setY(double y)
 {
     this->y = y;
 }
